+++ Reactive/Tests/ReactivePropertyTests.cs
@@ -1,12 +1,11 @@
 #nullable enable
 using System;
-using Blind.UiMvvm.ReactiveProperties;
-using Blind.UiMvvm.ReactiveProperties.Implementation;
+using Blind.Reactive.Implementation;
 using NUnit.Framework;
 using UnityEngine;
 using UnityEngine.TestTools;
 
-namespace Blind.UiMvvm.Tests
+namespace Blind.Reactive.Tests
 {
 	public class ReactivePropertyTests
 	{